Begin implementing PUT for questions
Having some issues when attempting to replace a multiple choice question with another multiple choice question. In SavvyRepository.UpdateQuestion, as soon as I attempt to access _ctx.Questions.Local, I get an InvalidOperationException:

The property 'ChoiceId' is part of the object's key information and cannot be modified.

The incoming question model gets parsed with all ChoiceId's being set to 0. The choices need to be attached to the context with a state of Added.